+++ src/widgets/dock_area/show/leaf.rs
@@ -1,12 +1,11 @@
+use egui::{
+    emath::TSTransform, epaint::TextShape, lerp, pos2, vec2, Align, Align2, Button, Color32,
+    CursorIcon, Frame, Id, Key, LayerId, Layout, NumExt, Order, Rect, Response, Rounding,
+    ScrollArea, Sense, Shape, Stroke, TextStyle, Ui, UiBuilder, Vec2, WidgetText,
+};
 use std::ops::RangeInclusive;
 
-use egui::emath::TSTransform;
-use egui::{
-    epaint::TextShape, lerp, pos2, vec2, Align, Align2, Button, CursorIcon, Frame, Id, Key,
-    LayerId, Layout, NumExt, Order, Rect, Response, Rounding, ScrollArea, Sense, Stroke, TextStyle,
-    Ui, UiBuilder, Vec2, WidgetText,
-};
-
+use crate::dock_area::tab_removal::TabRemoval;
 use crate::{
     dock_area::{
         drag_and_drop::{DragData, DragDropState, HoverData, TreeComponent},
@@ -18,7 +17,7 @@ use crate::{
 
 use crate::popup::popup_under_widget;
 
-impl<'tree, Tab> DockArea<'tree, Tab> {
+impl<Tab> DockArea<'_, Tab> {
     pub(super) fn show_leaf(
         &mut self,
         ui: &mut Ui,
@@ -28,6 +27,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         fade_style: Option<(&Style, f32)>,
     ) {
         assert!(self.dock_state[surface_index][node_index].is_leaf());
+        let collapsed = self.dock_state[surface_index][node_index].is_collapsed();
 
         let rect = self.dock_state[surface_index][node_index]
             .rect()
@@ -42,12 +42,16 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         ui.spacing_mut().item_spacing = Vec2::ZERO;
         ui.set_clip_rect(rect);
 
+        if self.dock_state[surface_index][node_index].tabs_count() == 0 {
+            return;
+        }
         let tabbar_rect = self.tab_bar(
             ui,
             state,
             (surface_index, node_index),
             tab_viewer,
             fade_style.map(|(style, _)| style),
+            collapsed,
         );
         self.tab_body(
             ui,
@@ -57,6 +61,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             spacing,
             tabbar_rect,
             fade_style,
+            collapsed,
         );
 
         let tabs = self.dock_state[surface_index][node_index]
@@ -77,6 +82,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         (surface_index, node_index): (SurfaceIndex, NodeIndex),
         tab_viewer: &mut impl TabViewer<Tab = Tab>,
         fade_style: Option<&Style>,
+        collapsed: bool,
     ) -> Rect {
         assert!(self.dock_state[surface_index][node_index].is_leaf());
 
@@ -92,13 +98,23 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         );
 
         let mut available_width = tabbar_outer_rect.width();
+        let scroll_bar_width = available_width;
         if available_width == 0.0 {
             return tabbar_outer_rect;
         }
 
-        // Reserve space for the add button at the end of the tab bar.
+        // Reserve space for the buttons at the ends of the tab bar.
+
         if self.show_add_buttons {
             available_width -= Style::TAB_ADD_BUTTON_SIZE;
+        }
+
+        if self.show_leaf_close_all_buttons {
+            available_width -= Style::TAB_CLOSE_ALL_BUTTON_SIZE;
+        }
+
+        if self.show_leaf_collapse_buttons {
+            available_width -= Style::TAB_COLLAPSE_BUTTON_SIZE;
         }
 
         let actual_width = {
@@ -108,7 +124,16 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             };
 
             let tabbar_inner_rect = Rect::from_min_size(
-                (tabbar_outer_rect.min - pos2(-*scroll, 0.0)).to_pos2(),
+                (tabbar_outer_rect.min - pos2(-*scroll, 0.0)
+                    + vec2(
+                        if self.show_leaf_collapse_buttons {
+                            Style::TAB_COLLAPSE_BUTTON_SIZE
+                        } else {
+                            0.0
+                        },
+                        0.0,
+                    ))
+                .to_pos2(),
                 vec2(tabbar_outer_rect.width(), tabbar_outer_rect.height()),
             );
 
@@ -121,6 +146,9 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
 
             let mut clip_rect = tabbar_outer_rect;
             clip_rect.set_width(available_width);
+            if self.show_leaf_collapse_buttons {
+                clip_rect = clip_rect.translate(vec2(Style::TAB_COLLAPSE_BUTTON_SIZE, 0.0));
+            }
             tabs_ui.set_clip_rect(clip_rect);
 
             // Desired size for tabs in "expanded" mode.
@@ -149,13 +177,17 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
                 (px, style.tab_bar.hline_color),
             );
 
-            // Add button at the end of the tab bar.
+            // Add button at the ends of the tab bar.
             if self.show_add_buttons {
                 let offset = match style.buttons.add_tab_align {
                     TabAddAlign::Left => {
                         (clip_rect.width() - tabs_ui.min_rect().width()).at_least(0.0)
                     }
                     TabAddAlign::Right => 0.0,
+                } + if self.show_leaf_close_all_buttons {
+                    Style::TAB_CLOSE_ALL_BUTTON_SIZE
+                } else {
+                    0.0
                 };
                 self.tab_plus(
                     ui,
@@ -168,6 +200,48 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
                 );
             }
 
+            if self.show_leaf_close_all_buttons {
+                // Current leaf contains non-closable tabs.
+                let disabled = if let Node::Leaf { tabs, .. } =
+                    &mut self.dock_state[surface_index][node_index]
+                {
+                    !tabs.iter_mut().all(|tab| tab_viewer.closeable(tab))
+                } else {
+                    unreachable!()
+                };
+
+                // Current window contains non-closable tabs.
+                let close_window_disabled = disabled
+                    || !self.dock_state[surface_index].iter_mut().all(|node| {
+                        if let Node::Leaf { tabs, .. } = node {
+                            tabs.iter_mut().all(|tab| tab_viewer.closeable(tab))
+                        } else {
+                            true
+                        }
+                    });
+
+                self.tab_close_all(
+                    ui,
+                    surface_index,
+                    node_index,
+                    tabbar_outer_rect,
+                    fade_style,
+                    disabled,
+                    close_window_disabled,
+                )
+            }
+
+            if self.show_leaf_collapse_buttons {
+                self.tab_collapse(
+                    ui,
+                    surface_index,
+                    node_index,
+                    tabbar_outer_rect,
+                    fade_style,
+                    collapsed,
+                )
+            }
+
             tabs_ui.min_rect().width()
         };
 
@@ -177,6 +251,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             (surface_index, node_index),
             actual_width,
             available_width,
+            scroll_bar_width,
             &tabbar_response,
             fade_style,
         );
@@ -241,7 +316,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             let (response, title_id) = if is_being_dragged {
                 let layer_id = LayerId::new(Order::Tooltip, id);
                 let response = tabs_ui
-                    .with_layer_id(layer_id, |ui| {
+                    .scope_builder(UiBuilder::new().layer_id(layer_id), |ui| {
                         self.tab_title(
                             ui,
                             &tab_style,
@@ -416,6 +491,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         }
     }
 
+    /// Draws the tab add button.
     #[allow(clippy::too_many_arguments)]
     fn tab_plus(
         &mut self,
@@ -488,6 +564,378 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         }
     }
 
+    /// Draws the close all button.
+    #[allow(clippy::too_many_arguments)]
+    #[allow(unused_assignments)]
+    fn tab_close_all(
+        &mut self,
+        ui: &mut Ui,
+        surface_index: SurfaceIndex,
+        node_index: NodeIndex,
+        tabbar_outer_rect: Rect,
+        fade_style: Option<&Style>,
+        disabled: bool,
+        close_window_disabled: bool,
+    ) {
+        let rect = Rect::from_min_max(
+            tabbar_outer_rect.right_top() - vec2(Style::TAB_CLOSE_ALL_BUTTON_SIZE, 0.0),
+            tabbar_outer_rect.right_bottom() - vec2(0.0, 2.0),
+        );
+
+        let ui = &mut ui.new_child(
+            UiBuilder::new()
+                .max_rect(rect)
+                .layout(Layout::left_to_right(Align::Center))
+                .id_salt((node_index, "tab_close_all")),
+        );
+
+        let (rect, mut response) = ui.allocate_exact_size(ui.available_size(), Sense::click());
+
+        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
+
+        // Whether we're on "secondary button mode" due to modifier keys
+        let on_secondary_button = self.is_on_secondary_button(surface_index, ui, &response);
+
+        let mut stroke_color = if disabled {
+            style.buttons.close_all_tabs_disabled_color
+        } else if response.hovered() || response.has_focus() {
+            if !(close_window_disabled && on_secondary_button) {
+                ui.painter().rect_filled(
+                    rect,
+                    Rounding::ZERO,
+                    style.buttons.close_all_tabs_bg_fill,
+                );
+            }
+            style.buttons.close_all_tabs_active_color
+        } else {
+            style.buttons.close_all_tabs_color
+        };
+
+        let mut close_all_rect = rect;
+
+        rect_set_size_centered(&mut close_all_rect, Vec2::splat(Style::TAB_CLOSE_ALL_SIZE));
+
+        if !disabled {
+            response = response.on_hover_cursor(CursorIcon::PointingHand);
+        }
+
+        if on_secondary_button {
+            // Close the entire window
+            if close_window_disabled {
+                stroke_color = style.buttons.close_all_tabs_disabled_color;
+                response = response
+                    .on_hover_cursor(CursorIcon::NotAllowed)
+                    .on_hover_text(
+                        self.dock_state
+                            .translations
+                            .leaf
+                            .close_all_button_disabled_tooltip
+                            .as_str(),
+                    );
+            }
+            Self::draw_close_window_symbol(ui, stroke_color, close_all_rect);
+        } else {
+            // Close all tabs in this leaf
+            if !disabled {
+                if !surface_index.is_main() && self.secondary_button_context_menu {
+                    response.context_menu(|ui| {
+                        ui.add_enabled_ui(!close_window_disabled, |ui| {
+                            if ui
+                                .button(&self.dock_state.translations.leaf.close_all_button)
+                                .on_disabled_hover_text(
+                                    self.dock_state
+                                        .translations
+                                        .leaf
+                                        .close_all_button_disabled_tooltip
+                                        .as_str(),
+                                )
+                                .clicked()
+                            {
+                                self.to_remove.push(TabRemoval::Window(surface_index));
+                            }
+                        });
+                    });
+                }
+            } else {
+                response = response
+                    .on_hover_cursor(CursorIcon::NotAllowed)
+                    .on_hover_text(
+                        self.dock_state
+                            .translations
+                            .leaf
+                            .close_button_disabled_tooltip
+                            .as_str(),
+                    );
+            }
+
+            if response.clicked() {
+                if on_secondary_button {
+                    if !close_window_disabled {
+                        self.to_remove.push(TabRemoval::Window(surface_index));
+                    }
+                } else if !disabled {
+                    self.to_remove.push((surface_index, node_index).into());
+                }
+            }
+
+            ui.painter().line_segment(
+                [close_all_rect.left_top(), close_all_rect.right_bottom()],
+                Stroke::new(1.0, stroke_color),
+            );
+            ui.painter().line_segment(
+                [close_all_rect.right_top(), close_all_rect.left_bottom()],
+                Stroke::new(1.0, stroke_color),
+            );
+        }
+
+        // Draw button left border.
+        ui.painter().vline(
+            rect.left(),
+            rect.y_range(),
+            Stroke::new(
+                ui.ctx().pixels_per_point().recip(),
+                style.buttons.close_all_tabs_border_color,
+            ),
+        );
+
+        if !disabled && !on_secondary_button {
+            response = self.show_tooltip_hints(surface_index, response);
+        }
+    }
+
+    /// Draws the collapse button.
+    fn tab_collapse(
+        &mut self,
+        ui: &mut Ui,
+        surface_index: SurfaceIndex,
+        node_index: NodeIndex,
+        tabbar_outer_rect: Rect,
+        fade_style: Option<&Style>,
+        collapsed: bool,
+    ) {
+        let rect = Rect::from_min_max(
+            tabbar_outer_rect.left_top(),
+            tabbar_outer_rect.left_bottom() + vec2(Style::TAB_COLLAPSE_BUTTON_SIZE, 0.0),
+        );
+
+        let ui = &mut ui.new_child(
+            UiBuilder::new()
+                .max_rect(rect)
+                .layout(Layout::left_to_right(Align::Center))
+                .id_salt((node_index, "tab_collapse")),
+        );
+
+        let (rect, mut response) = ui.allocate_exact_size(ui.available_size(), Sense::click());
+
+        response = response.on_hover_cursor(CursorIcon::PointingHand);
+
+        let style = fade_style.unwrap_or_else(|| self.style.as_ref().unwrap());
+
+        // Whether we're on "secondary button mode" due to modifier keys
+        let on_secondary_button = self.is_on_secondary_button(surface_index, ui, &response);
+
+        let color = if response.hovered() || response.has_focus() {
+            ui.painter()
+                .rect_filled(rect, Rounding::ZERO, style.buttons.collapse_tabs_bg_fill);
+            style.buttons.collapse_tabs_active_color
+        } else {
+            style.buttons.collapse_tabs_color
+        };
+
+        let mut arrow_rect = rect;
+        rect_set_size_centered(&mut arrow_rect, Vec2::splat(Style::TAB_COLLAPSE_ARROW_SIZE));
+
+        if on_secondary_button {
+            // Collapse the entire window
+            Self::draw_chevron_down(ui, style, color, arrow_rect);
+        } else {
+            // Draw arrow.
+            Self::draw_arrow(collapsed, ui, color, arrow_rect);
+        }
+
+        // Draw button right border.
+        ui.painter().vline(
+            rect.right(),
+            rect.y_range(),
+            Stroke::new(
+                ui.ctx().pixels_per_point().recip(),
+                style.buttons.collapse_tabs_border_color,
+            ),
+        );
+
+        if response.clicked() {
+            if on_secondary_button {
+                self.window_toggle_minimized(surface_index);
+            } else {
+                self.dock_state[surface_index][node_index].set_collapsed(!collapsed);
+                self.dock_state[surface_index].node_update_collapsed(node_index);
+                self.window_update_collapsed(surface_index, node_index);
+            }
+        }
+
+        if !surface_index.is_main() && self.secondary_button_context_menu {
+            response.context_menu(|ui| {
+                if ui
+                    .button(&self.dock_state.translations.leaf.minimize_button)
+                    .clicked()
+                {
+                    ui.close_menu();
+                    self.window_toggle_minimized(surface_index);
+                }
+            });
+        }
+
+        if !on_secondary_button {
+            self.show_tooltip_hints(surface_index, response);
+        }
+    }
+
+    fn show_tooltip_hints(&mut self, surface_index: SurfaceIndex, response: Response) -> Response {
+        if !surface_index.is_main()
+            && self.show_secondary_button_hint
+            && (self.secondary_button_context_menu || self.secondary_button_on_modifier)
+        {
+            let hint = if self.secondary_button_context_menu && self.secondary_button_on_modifier {
+                &self
+                    .dock_state
+                    .translations
+                    .leaf
+                    .minimize_button_modifier_menu_hint
+            } else if self.secondary_button_context_menu {
+                &self.dock_state.translations.leaf.minimize_button_menu_hint
+            } else {
+                &self
+                    .dock_state
+                    .translations
+                    .leaf
+                    .minimize_button_modifier_hint
+            };
+            return response.on_hover_text(hint);
+        }
+        response
+    }
+
+    fn is_on_secondary_button(
+        &self,
+        surface_index: SurfaceIndex,
+        ui: &mut Ui,
+        response: &Response,
+    ) -> bool {
+        !surface_index.is_main()
+            && self.secondary_button_on_modifier
+            && ui.input(|i| {
+                i.modifiers
+                    .matches_logically(self.secondary_button_modifiers)
+            })
+            && (response.hovered() || response.has_focus() || response.is_pointer_button_down_on())
+    }
+
+    fn draw_close_window_symbol(ui: &mut Ui, stroke_color: Color32, close_all_rect: Rect) {
+        ui.painter().add(Shape::line(
+            vec![
+                close_all_rect
+                    .right_center()
+                    .lerp(close_all_rect.right_bottom(), 0.5),
+                close_all_rect.right_bottom(),
+                close_all_rect.left_bottom(),
+                close_all_rect.left_top(),
+                close_all_rect
+                    .center_top()
+                    .lerp(close_all_rect.left_top(), 0.5),
+            ],
+            Stroke::new(1.0, stroke_color),
+        ));
+        ui.painter().line_segment(
+            [close_all_rect.center_top(), close_all_rect.right_center()],
+            Stroke::new(1.0, stroke_color),
+        );
+        ui.painter().line_segment(
+            [close_all_rect.center(), close_all_rect.right_top()],
+            Stroke::new(1.0, stroke_color),
+        );
+    }
+
+    fn draw_arrow(collapsed: bool, ui: &mut Ui, color: Color32, arrow_rect: Rect) {
+        ui.painter().add(Shape::convex_polygon(
+            if collapsed {
+                // Arrow pointing rightwards.
+                vec![
+                    arrow_rect.left_top(),
+                    arrow_rect.right_center(),
+                    arrow_rect.left_bottom(),
+                ]
+            } else {
+                // Arrow pointing downwards.
+                vec![
+                    arrow_rect.left_top(),
+                    arrow_rect.right_top(),
+                    arrow_rect.center_bottom(),
+                ]
+            },
+            color,
+            Stroke::NONE,
+        ));
+    }
+
+    fn draw_chevron_down(ui: &mut Ui, style: &Style, color: Color32, arrow_rect: Rect) {
+        ui.painter().add(Shape::convex_polygon(
+            // Arrow pointing downwards.
+            vec![
+                arrow_rect.left_top(),
+                arrow_rect.right_top(),
+                arrow_rect.center(),
+            ],
+            color,
+            Stroke::NONE,
+        ));
+
+        // Chevron pointing downwards.
+        ui.painter().add(Shape::convex_polygon(
+            vec![
+                arrow_rect.left_center(),
+                arrow_rect.right_center(),
+                arrow_rect.center_bottom(),
+            ],
+            color,
+            Stroke::NONE,
+        ));
+        let color = style.buttons.minimize_window_bg_fill;
+        ui.painter().add(Shape::convex_polygon(
+            vec![
+                arrow_rect
+                    .left_center()
+                    .lerp(arrow_rect.right_center(), 0.25),
+                arrow_rect
+                    .left_center()
+                    .lerp(arrow_rect.right_center(), 0.75),
+                arrow_rect.center().lerp(arrow_rect.center_bottom(), 0.5),
+            ],
+            color,
+            Stroke::NONE,
+        ));
+    }
+
+    /// Updates the collapsed state of the node and its parents.
+    fn window_update_collapsed(&mut self, surface_index: SurfaceIndex, node_index: NodeIndex) {
+        let surface = &mut self.dock_state[surface_index];
+        let collapsed = surface[node_index].is_collapsed();
+        if !collapsed {
+            if let Some(window_state) = self.dock_state.get_window_state_mut(surface_index) {
+                window_state.set_new(true);
+            }
+        } else if surface.root_node().is_some_and(|root| root.is_collapsed()) {
+            let root_index = NodeIndex::root();
+            let surface_height = if surface.root_node().is_some() {
+                surface[root_index].rect().unwrap().height()
+            } else {
+                0.0
+            };
+            if let Some(window_state) = self.dock_state.get_window_state_mut(surface_index) {
+                window_state.set_expanded_height(surface_height);
+            }
+        }
+    }
+
     /// * `active` means "the tab that is opened in the parent panel".
     /// * `focused` means "the tab that was last interacted with".
     ///
@@ -526,7 +974,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         let (_, tab_rect) = ui.allocate_space(vec2(tab_width, ui.available_height()));
         let mut response = ui.interact(tab_rect, id, Sense::click_and_drag());
         if ui.ctx().dragged_id().is_none() && self.draggable_tabs {
-            response = response.on_hover_cursor(CursorIcon::PointingHand);
+            response = response.on_hover_cursor(CursorIcon::Grab);
         }
 
         let tab_style = if focused || is_being_dragged {
@@ -634,6 +1082,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         (surface_index, node_index): (SurfaceIndex, NodeIndex),
         actual_width: f32,
         available_width: f32,
+        scroll_bar_width: f32,
         tabbar_response: &Response,
         fade_style: Option<&Style>,
     ) {
@@ -652,7 +1101,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
                 // Draw scroll bar
                 let bar_height = 7.5;
                 let (scroll_bar_rect, _scroll_bar_response) = ui.allocate_exact_size(
-                    vec2(available_width, bar_height),
+                    vec2(scroll_bar_width, bar_height),
                     Sense::click_and_drag(),
                 );
 
@@ -722,6 +1171,7 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
         spacing: Vec2,
         tabbar_rect: Rect,
         fade: Option<(&Style, f32)>,
+        collapsed: bool,
     ) {
         let (body_rect, _body_response) =
             ui.allocate_exact_size(ui.available_size_before_wrap(), Sense::hover());
@@ -737,74 +1187,78 @@ impl<'tree, Tab> DockArea<'tree, Tab> {
             unreachable!();
         };
 
-        if let Some(tab) = tabs.get_mut(active.0) {
-            *viewport = body_rect;
+        if !collapsed {
+            if let Some(tab) = tabs.get_mut(active.0) {
+                *viewport = body_rect;
 
-            if ui.input(|i| i.pointer.any_click()) {
-                if let Some(pos) = state.last_hover_pos {
-                    if body_rect.contains(pos) && Some(ui.layer_id()) == ui.ctx().layer_id_at(pos) {
-                        self.new_focused = Some((surface_index, node_index));
+                if ui.input(|i| i.pointer.any_click()) {
+                    if let Some(pos) = state.last_hover_pos {
+                        if body_rect.contains(pos)
+                            && Some(ui.layer_id()) == ui.ctx().layer_id_at(pos)
+                        {
+                            self.new_focused = Some((surface_index, node_index));
+                        }
                     }
                 }
-            }
 
-            let (style, fade_factor) = fade.unwrap_or_else(|| (self.style.as_ref().unwrap(), 1.0));
-            let tabs_styles = tab_viewer.tab_style_override(tab, &style.tab);
+                let (style, fade_factor) =
+                    fade.unwrap_or_else(|| (self.style.as_ref().unwrap(), 1.0));
+                let tabs_styles = tab_viewer.tab_style_override(tab, &style.tab);
 
-            let tabs_style = tabs_styles.as_ref().unwrap_or(&style.tab);
+                let tabs_style = tabs_styles.as_ref().unwrap_or(&style.tab);
 
-            if tab_viewer.clear_background(tab) {
-                ui.painter().rect_filled(
-                    body_rect,
-                    tabs_style.tab_body.rounding,
-                    tabs_style.tab_body.bg_fill,
+                if tab_viewer.clear_background(tab) {
+                    ui.painter().rect_filled(
+                        body_rect,
+                        tabs_style.tab_body.rounding,
+                        tabs_style.tab_body.bg_fill,
+                    );
+                }
+
+                // Construct a new ui with the correct tab id.
+                //
+                // We are forced to use `Ui::new` because other methods (eg: push_id) always mix
+                // the provided id with their own which would cause tabs to change id when moved
+                // from node to node.
+                let id = self.id.with(tab_viewer.id(tab));
+                ui.ctx().check_for_id_clash(id, body_rect, "a tab with id");
+                let ui = &mut Ui::new(
+                    ui.ctx().clone(),
+                    id,
+                    UiBuilder::new().max_rect(body_rect).layer_id(ui.layer_id()),
                 );
+                ui.set_clip_rect(Rect::from_min_max(ui.cursor().min, ui.clip_rect().max));
+
+                // Use initial spacing for ui.
+                ui.spacing_mut().item_spacing = spacing;
+
+                // Offset the background rectangle up to hide the top border behind the clip rect.
+                // To avoid anti-aliasing lines when the stroke width is not divisible by two, we
+                // need to calculate the effective anti-aliased stroke width.
+                let effective_stroke_width = (tabs_style.tab_body.stroke.width / 2.0).ceil() * 2.0;
+                let tab_body_rect = Rect::from_min_max(
+                    ui.clip_rect().min - vec2(0.0, effective_stroke_width),
+                    ui.clip_rect().max,
+                );
+                ui.painter().rect_stroke(
+                    rect_stroke_box(tab_body_rect, tabs_style.tab_body.stroke.width),
+                    tabs_style.tab_body.rounding,
+                    tabs_style.tab_body.stroke,
+                );
+
+                ScrollArea::new(tab_viewer.scroll_bars(tab)).show(ui, |ui| {
+                    Frame::none()
+                        .inner_margin(tabs_style.tab_body.inner_margin)
+                        .show(ui, |ui| {
+                            if fade_factor != 1.0 {
+                                fade_visuals(ui.visuals_mut(), fade_factor);
+                            }
+                            let available_rect = ui.available_rect_before_wrap();
+                            ui.expand_to_include_rect(available_rect);
+                            tab_viewer.ui(ui, tab);
+                        });
+                });
             }
-
-            // Construct a new ui with the correct tab id.
-            //
-            // We are forced to use `Ui::new` because other methods (eg: push_id) always mix
-            // the provided id with their own which would cause tabs to change id when moved
-            // from node to node.
-            let id = self.id.with(tab_viewer.id(tab));
-            ui.ctx().check_for_id_clash(id, body_rect, "a tab with id");
-            let ui = &mut Ui::new(
-                ui.ctx().clone(),
-                ui.layer_id(),
-                id,
-                UiBuilder::new().max_rect(body_rect),
-            );
-            ui.set_clip_rect(Rect::from_min_max(ui.cursor().min, ui.clip_rect().max));
-
-            // Use initial spacing for ui.
-            ui.spacing_mut().item_spacing = spacing;
-
-            // Offset the background rectangle up to hide the top border behind the clip rect.
-            // To avoid anti-aliasing lines when the stroke width is not divisible by two, we
-            // need to calculate the effective anti-aliased stroke width.
-            let effective_stroke_width = (tabs_style.tab_body.stroke.width / 2.0).ceil() * 2.0;
-            let tab_body_rect = Rect::from_min_max(
-                ui.clip_rect().min - vec2(0.0, effective_stroke_width),
-                ui.clip_rect().max,
-            );
-            ui.painter().rect_stroke(
-                rect_stroke_box(tab_body_rect, tabs_style.tab_body.stroke.width),
-                tabs_style.tab_body.rounding,
-                tabs_style.tab_body.stroke,
-            );
-
-            ScrollArea::new(tab_viewer.scroll_bars(tab)).show(ui, |ui| {
-                Frame::none()
-                    .inner_margin(tabs_style.tab_body.inner_margin)
-                    .show(ui, |ui| {
-                        if fade_factor != 1.0 {
-                            fade_visuals(ui.visuals_mut(), fade_factor);
-                        }
-                        let available_rect = ui.available_rect_before_wrap();
-                        ui.expand_to_include_rect(available_rect);
-                        tab_viewer.ui(ui, tab);
-                    });
-            });
         }
 
         // change hover destination